backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
pandas>=2.2
numpy>=1.26
openpyxl>=3.1
scikit-learn>=1.4
scipy>=1.12
anthropic>=0.40
python-dotenv>=1.0
